LP-20: exceptions

+++ src/main/java/com/nlitvins/web_application/domain/exception/UserNotFoundException.java
@@ -1,7 +1,7 @@
 package com.nlitvins.web_application.domain.exception;
 
 public class UserNotFoundException extends RuntimeException {
-    public UserNotFoundException(Long id) {
+    public UserNotFoundException(int id) {
         super("User not found with id: " + id);
     }
 }
